requirements.txt
Django>=4.1,<4.2
wagtail>=4.1,<4.2
pip==22.3.1
fontawesomefree>=6.2,<6.3
black
flake8==6.0.0
isort
djlint
psycopg2-binary
python-dotenv
django-crispy-forms
crispy-tailwind
django-allauth==0.51.0
django-debug-toolbar==3.8.1
django-extensions==3.2.1
django-countries
whitenoise==6.2.0
django-storages==1.13.1
wagtail-localize==1.3.3
wagtail-django-recaptcha==1.0